requirements.txt
torch>=1.8.0
torchvision>=0.23.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0